Simplify local dev env setup (#192)
* Simplify local dev env setup

* fmt

* Fix some deps

* More dep fixes

* fmt

* Change clean bahaviour
