+++ azero/package.json
@@ -5,7 +5,6 @@
     "license": "MIT",
     "description": "azero most bridge",
     "scripts": {
-        "watch": "gulp default",
         "compile": "typechain-compiler --release --toolchain nightly-2024-02-16",
         "predrink-tests": "npm config set script-shell \"/bin/bash\"; ./scripts/predrink_tests.sh",
         "drink-tests": "npm config set script-shell \"/bin/bash\"; ./scripts/drink_tests.sh",
@@ -32,8 +31,6 @@
     },
     "devDependencies": {
         "ts-node": "^10.9.1",
-        "typescript": "^5.1.3",
-        "gulp": "^5.0.0",
-        "gulp-shell": "^0.8.0"
+        "typescript": "^5.1.3"
     }
 }
