--- eth/contracts/gnosis-1.3.0/base/FallbackManager.sol
@@ -1,53 +1,0 @@
-// SPDX-License-Identifier: LGPL-3.0-only
-pragma solidity >=0.7.0 <0.9.0;
-
-import "../common/SelfAuthorized.sol";
-
-/// @title Fallback Manager - A contract that manages fallback calls made to this contract
-/// @author Richard Meissner - <richard@gnosis.pm>
-contract FallbackManager is SelfAuthorized {
-    event ChangedFallbackHandler(address handler);
-
-    // keccak256("fallback_manager.handler.address")
-    bytes32 internal constant FALLBACK_HANDLER_STORAGE_SLOT = 0x6c9a6c4a39284e37ed1cf53d337577d14212a4870fb976a4366c693b939918d5;
-
-    function internalSetFallbackHandler(address handler) internal {
-        bytes32 slot = FALLBACK_HANDLER_STORAGE_SLOT;
-        // solhint-disable-next-line no-inline-assembly
-        assembly {
-            sstore(slot, handler)
-        }
-    }
-
-    /// @dev Allows to add a contract to handle fallback calls.
-    ///      Only fallback calls without value and with data will be forwarded.
-    ///      This can only be done via a Safe transaction.
-    /// @param handler contract to handle fallback calls.
-    function setFallbackHandler(address handler) public authorized {
-        internalSetFallbackHandler(handler);
-        emit ChangedFallbackHandler(handler);
-    }
-
-    // solhint-disable-next-line payable-fallback,no-complex-fallback
-    fallback() external {
-        bytes32 slot = FALLBACK_HANDLER_STORAGE_SLOT;
-        // solhint-disable-next-line no-inline-assembly
-        assembly {
-            let handler := sload(slot)
-            if iszero(handler) {
-                return(0, 0)
-            }
-            calldatacopy(0, 0, calldatasize())
-            // The msg.sender address is shifted to the left by 12 bytes to remove the padding
-            // Then the address without padding is stored right after the calldata
-            mstore(calldatasize(), shl(96, caller()))
-            // Add 20 bytes for the address appended add the end
-            let success := call(gas(), handler, 0, 0, add(calldatasize(), 20), 0, 0)
-            returndatacopy(0, 0, returndatasize())
-            if iszero(success) {
-                revert(0, returndatasize())
-            }
-            return(0, returndatasize())
-        }
-    }
-}
