--- eth/contracts/gnosis-1.3.0/common/Enum.sol
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: LGPL-3.0-only
-pragma solidity >=0.7.0 <0.9.0;
-
-/// @title Enum - Collection of enums
-/// @author Richard Meissner - <richard@gnosis.pm>
-contract Enum {
-    enum Operation {Call, DelegateCall}
-}
