--- eth/contracts/gnosis-1.3.0/common/SelfAuthorized.sol
@@ -1,16 +1,0 @@
-// SPDX-License-Identifier: LGPL-3.0-only
-pragma solidity >=0.7.0 <0.9.0;
-
-/// @title SelfAuthorized - authorizes current contract to perform actions
-/// @author Richard Meissner - <richard@gnosis.pm>
-contract SelfAuthorized {
-    function requireSelfCall() private view {
-        require(msg.sender == address(this), "GS031");
-    }
-
-    modifier authorized() {
-        // This is a function call as it minimized the bytecode size
-        requireSelfCall();
-        _;
-    }
-}
