--- eth/contracts/gnosis-1.3.0/common/Singleton.sol
@@ -1,11 +1,0 @@
-// SPDX-License-Identifier: LGPL-3.0-only
-pragma solidity >=0.7.0 <0.9.0;
-
-/// @title Singleton - Base for singleton contracts (should always be first super contract)
-///         This contract is tightly coupled to our proxy contract (see `proxies/GnosisSafeProxy.sol`)
-/// @author Richard Meissner - <richard@gnosis.io>
-contract Singleton {
-    // singleton always needs to be first declared variable, to ensure that it is at the same location as in the Proxy contract.
-    // It should also always be ensured that the address is stored alone (uses a full word)
-    address private singleton;
-}
