--- eth/contracts/gnosis-1.3.0/interfaces/ERC777TokensRecipient.sol
@@ -1,13 +1,0 @@
-// SPDX-License-Identifier: LGPL-3.0-only
-pragma solidity >=0.7.0 <0.9.0;
-
-interface ERC777TokensRecipient {
-    function tokensReceived(
-        address operator,
-        address from,
-        address to,
-        uint256 amount,
-        bytes calldata data,
-        bytes calldata operatorData
-    ) external;
-}
