--- eth/contracts/gnosis-1.3.0/proxies/IProxyCreationCallback.sol
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: LGPL-3.0-only
-pragma solidity >=0.7.0 <0.9.0;
-import "./GnosisSafeProxy.sol";
-
-interface IProxyCreationCallback {
-    function proxyCreated(
-        GnosisSafeProxy proxy,
-        address _singleton,
-        bytes calldata initializer,
-        uint256 saltNonce
-    ) external;
-}
