+++ eth/package.json
@@ -4,9 +4,6 @@
     "description": "evm most bridge contracts",
     "author": "fbielejec",
     "license": "MIT",
-    "scripts": {
-        "watch": "gulp default"
-    },
     "dependencies": {
         "@openzeppelin/contracts": "=5.0.2",
         "@openzeppelin/contracts-upgradeable": "=5.0.2"
@@ -18,12 +15,8 @@
         "@nomiclabs/hardhat-truffle5": "^2.0.7",
         "@openzeppelin/hardhat-upgrades": "^2.5.0",
         "@polkadot/api": "^10.11.2",
-        "@safe-global/protocol-kit": "^3.0.0",
-        "@safe-global/safe-core-sdk-types": "^4.0.0",
         "child_process": "^1.0.2",
         "dotenv": "^16.3.1",
-        "gulp": "^5.0.0",
-        "gulp-shell": "^0.8.0",
         "hardhat": "^2.19.0",
         "prettier": "^3.2.5",
         "prettier-plugin-solidity": "^1.3.1",
